+++ crates/bin/starknet-sierra-compile/Cargo.toml
@@ -12,6 +12,7 @@ clap.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 
-cairo-lang-sierra = { path = "../../cairo-lang-sierra", version = "2.1.0-rc0" }
-cairo-lang-starknet = { path = "../../cairo-lang-starknet", version = "2.1.0-rc0" }
-cairo-lang-utils = { path = "../../cairo-lang-utils", version = "2.1.0-rc0" }
+cairo-lang-sierra = { path = "../../cairo-lang-sierra", version = "2.1.0-rc0", default-features = false }
+cairo-lang-starknet = { path = "../../cairo-lang-starknet", version = "2.1.0-rc0", default-features = false }
+cairo-lang-utils = { path = "../../cairo-lang-utils", version = "2.1.0-rc0", default-features = false }
+cairo-lang-casm-contract-class = { path = "../../cairo-lang-casm-contract-class", version = "2.1.0-rc0", default-features = false }
