+++ crates/cairo-lang-casm/Cargo.toml
@@ -7,15 +7,20 @@ license-file.workspace = true
 description = "Cairo assembly encoding."
 
 [dependencies]
-cairo-lang-utils = { path = "../cairo-lang-utils", version = "2.1.0-rc0" }
+cairo-lang-utils = { path = "../cairo-lang-utils", version = "2.1.0-rc0", default-features = false, features = [
+    "parity-scale-codec",
+] }
 indoc.workspace = true
 num-bigint.workspace = true
 num-traits.workspace = true
-parity-scale-codec.workspace = true
-parity-scale-codec-derive.workspace = true
-schemars = { workspace = true, features = ["preserve_order"] }
+parity-scale-codec = { workspace = true, optional = true, features = [
+    "derive",
+] }
 serde.workspace = true
-thiserror.workspace = true
+hashbrown.workspace = true
+thiserror-no-std.workspace = true
+# Optional feature
+schemars = { workspace = true, optional = true }
 
 [dev-dependencies]
 env_logger.workspace = true
@@ -23,3 +28,8 @@ itertools.workspace = true
 pretty_assertions.workspace = true
 test-case.workspace = true
 test-log.workspace = true
+
+[features]
+default = ["std"]
+std = ["cairo-lang-utils/std"]
+schemars = ["std", "dep:schemars", "cairo-lang-utils/schemars"]
