+++ crates/cairo-lang-casm/src/encoder_test.rs
@@ -1,3 +1,6 @@
+#[cfg(not(feature = "std"))]
+use alloc::{vec, vec::Vec};
+
 use num_bigint::BigInt;
 use pretty_assertions::assert_eq;
 use test_case::test_case;
