+++ crates/cairo-lang-casm/src/lib.rs
@@ -1,4 +1,8 @@
 //! Cairo assembly representation, formatting and construction utilities.
+#![cfg_attr(not(feature = "std"), no_std)]
+
+#[cfg(not(feature = "std"))]
+extern crate alloc;
 
 pub mod ap_change;
 pub mod assembler;
