+++ crates/cairo-lang-casm/src/operand.rs
@@ -1,25 +1,25 @@
-use std::fmt::Display;
+use core::fmt::Display;
 
 use cairo_lang_utils::bigint::BigIntAsHex;
-use parity_scale_codec_derive::{Decode, Encode};
-use schemars::JsonSchema;
+#[cfg(feature = "parity-scale-codec")]
+use parity_scale_codec::{Decode, Encode};
 use serde::{Deserialize, Serialize};
 
 #[cfg(test)]
 #[path = "operand_test.rs"]
 mod test;
 
-#[derive(
-    Copy, Clone, Debug, Hash, PartialEq, Eq, Serialize, Deserialize, Encode, Decode, JsonSchema,
-)]
+#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq, Serialize, Deserialize)]
+#[cfg_attr(feature = "schemars", derive(schemars::JsonSchema))]
+#[cfg_attr(feature = "parity-scale-codec", derive(Encode, Decode))]
 pub enum Register {
-    #[codec(index = 0)]
+    #[cfg_attr(feature = "parity-scale-codec", codec(index = 0))]
     AP,
-    #[codec(index = 1)]
+    #[cfg_attr(feature = "parity-scale-codec", codec(index = 1))]
     FP,
 }
 impl Display for Register {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         match self {
             Register::AP => write!(f, "ap"),
             Register::FP => write!(f, "fp"),
@@ -28,19 +28,21 @@ impl Display for Register {
 }
 
 // Represents the rhs operand of an assert equal InstructionBody.
-#[derive(Serialize, Deserialize, Clone, Debug, Eq, PartialEq, Encode, Decode, JsonSchema)]
+#[derive(Serialize, Deserialize, Clone, Debug, Eq, PartialEq)]
+#[cfg_attr(feature = "schemars", derive(schemars::JsonSchema))]
+#[cfg_attr(feature = "parity-scale-codec", derive(Encode, Decode))]
 pub enum ResOperand {
-    #[codec(index = 0)]
+    #[cfg_attr(feature = "parity-scale-codec", codec(index = 0))]
     Deref(CellRef),
-    #[codec(index = 1)]
+    #[cfg_attr(feature = "parity-scale-codec", codec(index = 1))]
     DoubleDeref(CellRef, i16),
-    #[codec(index = 2)]
+    #[cfg_attr(feature = "parity-scale-codec", codec(index = 2))]
     Immediate(BigIntAsHex),
-    #[codec(index = 3)]
+    #[cfg_attr(feature = "parity-scale-codec", codec(index = 3))]
     BinOp(BinOpOperand),
 }
 impl Display for ResOperand {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         match self {
             ResOperand::Deref(operand) => write!(f, "{operand}"),
             ResOperand::DoubleDeref(operand, offset) => write!(f, "[{operand} + {offset}]"),
@@ -65,13 +67,16 @@ impl<T: Into<BigIntAsHex>> From<T> for ResOperand {
 }
 
 /// Represents an operand of the form [reg + offset].
-#[derive(Serialize, Deserialize, Copy, Clone, Debug, Eq, PartialEq, Encode, Decode, JsonSchema)]
+#[derive(Serialize, Deserialize, Copy, Clone, Debug, Eq, PartialEq)]
+#[cfg_attr(feature = "schemars", derive(schemars::JsonSchema))]
+#[cfg_attr(feature = "parity-scale-codec", derive(Encode, Decode))]
+
 pub struct CellRef {
     pub register: Register,
     pub offset: i16,
 }
 impl Display for CellRef {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         write!(f, "[{} + {}]", self.register, self.offset)
     }
 }
@@ -81,15 +86,17 @@ pub fn ap_cell_ref(offset: i16) -> CellRef {
     CellRef { register: Register::AP, offset }
 }
 
-#[derive(Serialize, Deserialize, Clone, Debug, Eq, PartialEq, Encode, Decode, JsonSchema)]
+#[derive(Serialize, Deserialize, Clone, Debug, Eq, PartialEq)]
+#[cfg_attr(feature = "schemars", derive(schemars::JsonSchema))]
+#[cfg_attr(feature = "parity-scale-codec", derive(Encode, Decode))]
 pub enum DerefOrImmediate {
-    #[codec(index = 0)]
+    #[cfg_attr(feature = "parity-scale-codec", codec(index = 0))]
     Deref(CellRef),
-    #[codec(index = 1)]
+    #[cfg_attr(feature = "parity-scale-codec", codec(index = 1))]
     Immediate(BigIntAsHex),
 }
 impl Display for DerefOrImmediate {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         match self {
             DerefOrImmediate::Deref(operand) => write!(f, "{operand}"),
             DerefOrImmediate::Immediate(operand) => write!(f, "{}", operand.value),
@@ -107,15 +114,17 @@ impl From<CellRef> for DerefOrImmediate {
     }
 }
 
-#[derive(Serialize, Deserialize, Clone, Debug, Eq, PartialEq, Encode, Decode, JsonSchema)]
+#[derive(Serialize, Deserialize, Clone, Debug, Eq, PartialEq)]
+#[cfg_attr(feature = "schemars", derive(schemars::JsonSchema))]
+#[cfg_attr(feature = "parity-scale-codec", derive(Encode, Decode))]
 pub enum Operation {
-    #[codec(index = 0)]
+    #[cfg_attr(feature = "parity-scale-codec", codec(index = 0))]
     Add,
-    #[codec(index = 1)]
+    #[cfg_attr(feature = "parity-scale-codec", codec(index = 1))]
     Mul,
 }
 impl Display for Operation {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         match self {
             Operation::Add => write!(f, "+"),
             Operation::Mul => write!(f, "*"),
@@ -123,14 +132,16 @@ impl Display for Operation {
     }
 }
 
-#[derive(Serialize, Deserialize, Clone, Debug, Eq, PartialEq, Encode, Decode, JsonSchema)]
+#[derive(Serialize, Deserialize, Clone, Debug, Eq, PartialEq)]
+#[cfg_attr(feature = "schemars", derive(schemars::JsonSchema))]
+#[cfg_attr(feature = "parity-scale-codec", derive(Encode, Decode))]
 pub struct BinOpOperand {
     pub op: Operation,
     pub a: CellRef,
     pub b: DerefOrImmediate,
 }
 impl Display for BinOpOperand {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         write!(f, "{} {} {}", self.a, self.op, self.b)
     }
 }
