+++ crates/cairo-lang-runner/src/lib.rs
@@ -40,7 +40,6 @@ use num_traits::ToPrimitive;
 use thiserror::Error;
 
 pub mod casm_run;
-pub mod short_string;
 
 #[derive(Debug, Error)]
 pub enum RunnerError {
