+++ crates/cairo-lang-starknet/Cargo.toml
@@ -25,6 +25,7 @@ cairo-lang-sierra-generator = { path = "../cairo-lang-sierra-generator", version
 cairo-lang-sierra-to-casm = { path = "../cairo-lang-sierra-to-casm", version = "2.1.0-rc0" }
 cairo-lang-syntax = { path = "../cairo-lang-syntax", version = "2.1.0-rc0" }
 cairo-lang-utils = { path = "../cairo-lang-utils", version = "2.1.0-rc0" }
+cairo-lang-casm-contract-class = { path = "../cairo-lang-casm-contract-class", version = "2.1.0-rc0" }
 convert_case.workspace = true
 genco.workspace = true
 indoc.workspace = true
