+++ crates/cairo-lang-starknet/src/contract_class_into_casm_contract_class_test.rs
@@ -4,8 +4,10 @@ use cairo_felt::Felt252;
 use cairo_lang_test_utils::compare_contents_or_fix_with_path;
 use test_case::test_case;
 
-use crate::casm_contract_class::{BigUintAsHex, CasmContractClass, StarknetSierraCompilationError};
 use crate::contract_class::ContractClass;
+use crate::contract_class_into_casm_contract_class::{
+    BigUintAsHex, StarknetSierraCompilationError,
+};
 use crate::test_utils::{get_example_file_path, get_test_contract};
 
 /// Tests that the casm compiled from <test_case>.cairo is the same as in <test_case>.casm.json.
@@ -19,8 +21,7 @@ use crate::test_utils::{get_example_file_path, get_test_contract};
 fn test_casm_contract_from_contract_class(example_file_name: &str) {
     let contract_class = get_test_contract(format!("{example_file_name}.cairo").as_str());
     let add_pythonic_hints = true;
-    let casm_contract =
-        CasmContractClass::from_contract_class(contract_class, add_pythonic_hints).unwrap();
+    let casm_contract = contract_class.into_casm_contract_class(add_pythonic_hints).unwrap();
 
     compare_contents_or_fix_with_path(
         &get_example_file_path(format!("{example_file_name}.casm.json").as_str()),
@@ -38,7 +39,7 @@ fn test_casm_contract_from_contract_class_failure(example_file_name: &str) {
 
     let add_pythonic_hints = false;
     assert_eq!(
-        CasmContractClass::from_contract_class(contract_class, add_pythonic_hints),
+        contract_class.into_casm_contract_class(add_pythonic_hints),
         Err(StarknetSierraCompilationError::ValueOutOfRange)
     );
 }
