+++ crates/cairo-lang-syntax/Cargo.toml
@@ -14,7 +14,6 @@ num-bigint.workspace = true
 num-traits.workspace = true
 salsa.workspace = true
 smol_str.workspace = true
-thiserror.workspace = true
 unescaper.workspace = true
 
 [dev-dependencies]
