+++ crates/cairo-lang-test-runner/Cargo.toml
@@ -33,4 +33,3 @@ num-bigint.workspace = true
 num-traits.workspace = true
 rayon.workspace = true
 salsa.workspace = true
-thiserror.workspace = true
