+++ crates/cairo-lang-utils/Cargo.toml
@@ -12,12 +12,18 @@ itertools.workspace = true
 num-bigint.workspace = true
 num-integer.workspace = true
 num-traits.workspace = true
-parity-scale-codec.workspace = true
-schemars = { workspace = true, features = ["preserve_order"] }
+parity-scale-codec = { workspace = true, optional = true, features = [
+    "derive",
+] }
 serde.workspace = true
+cairo-felt = { workspace = true, features = ["alloc"] }
+hashbrown.workspace = true
 
-# Optional
+# Optional features
+schemars = { workspace = true, optional = true }
 env_logger = { workspace = true, optional = true }
+
+# Optional dependencies
 time = { workspace = true, optional = true }
 log = { workspace = true, optional = true }
 
@@ -28,5 +34,15 @@ test-log.workspace = true
 env_logger.workspace = true
 
 [features]
+default = ["std"]
+std = [
+    "indexmap/std",
+    "itertools/use_std",
+    "num-bigint/std",
+    "num-integer/std",
+    "num-traits/std",
+    "serde/std",
+    "cairo-felt/std",
+]
+env_logger = ["std", "dep:env_logger", "dep:time", "dep:log"]
 testing = []
-env_logger = ["dep:env_logger", "dep:time", "dep:log"]
