+++ crates/cairo-lang-utils/src/lib.rs
@@ -1,5 +1,8 @@
 //! Cairo utilities.
-use std::fmt;
+#![cfg_attr(not(feature = "std"), no_std)]
+
+#[cfg(not(feature = "std"))]
+extern crate alloc;
 
 pub mod bigint;
 pub mod casts;
@@ -11,8 +14,13 @@ pub mod iterators;
 pub mod logging;
 pub mod ordered_hash_map;
 pub mod ordered_hash_set;
+pub mod short_string;
 pub mod unordered_hash_map;
 pub mod unordered_hash_set;
+
+#[cfg(not(feature = "std"))]
+use alloc::boxed::Box;
+pub use core::{fmt, mem};
 
 /// Similar to From / TryFrom, but returns an option.
 pub trait OptionFrom<T>
@@ -22,7 +30,7 @@ where
     fn option_from(other: T) -> Option<Self>;
 }
 
-pub fn write_comma_separated<Iter: IntoIterator<Item = V>, V: std::fmt::Display>(
+pub fn write_comma_separated<Iter: IntoIterator<Item = V>, V: fmt::Display>(
     f: &mut fmt::Formatter<'_>,
     values: Iter,
 ) -> fmt::Result {
@@ -82,7 +90,7 @@ impl<T, E> ResultHelper<E> for Result<T, E> {
 pub fn borrow_as_box<T: Default, R, F: FnOnce(Box<T>) -> (R, Box<T>)>(ptr: &mut T, f: F) -> R {
     // TODO(spapini): Consider replacing take with something the leaves the memory dangling, instead
     // of filling with default().
-    let (res, boxed) = f(Box::new(std::mem::take(ptr)));
+    let (res, boxed) = f(Box::new(mem::take(ptr)));
     *ptr = *boxed;
     res
 }
@@ -113,10 +121,9 @@ macro_rules! define_short_id {
         impl<T: ?Sized + cairo_lang_utils::Upcast<dyn $db + 'static>>
             cairo_lang_debug::DebugWithDb<T> for $short_id
         {
-            fn fmt(&self, f: &mut std::fmt::Formatter<'_>, db: &T) -> std::fmt::Result {
-                use std::fmt::Debug;
-
+            fn fmt(&self, f: &mut $crate::fmt::Formatter<'_>, db: &T) -> $crate::fmt::Result {
                 use cairo_lang_debug::helper::Fallback;
+                use $crate::fmt::Debug;
                 let db = db.upcast();
                 cairo_lang_debug::helper::HelperDebug::<$long_id, dyn $db>::helper_debug(
                     &db.$lookup(*self),
